+++ template/frontend/ambev.developer-evaluation/src/app/products/components/product-list/product-list.component.ts
@@ -2,8 +2,7 @@ import { Component, inject, OnInit } from '@angular/core';
 import { Product } from '../../models/product.model';
 import { ProductService } from '../../services/product.service';
 import { CommonModule } from '@angular/common';
-import { Router, RouterOutlet } from '@angular/router';
-import { ConfirmationDialogComponent } from '../../../core/shared/components/confirmation-dialog/confirmation-dialog.component';
+import { Router } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
 import { MatIconModule } from '@angular/material/icon';
 import { MaterialModule } from '../../../core/shared/modules/material.module';
@@ -11,7 +10,7 @@ import { MaterialModule } from '../../../core/shared/modules/material.module';
 @Component({
   selector: 'app-product-list',
   standalone: true,
-  imports: [CommonModule, RouterOutlet, MaterialModule, MatIconModule],
+  imports: [CommonModule, MaterialModule, MatIconModule],
   templateUrl: './product-list.component.html',
   styleUrl: './product-list.component.scss',
 })
@@ -44,25 +43,16 @@ export class ProductListComponent implements OnInit {
     this.router.navigate(['products/add']);
   }
 
-  deleteProduct(id: number) {
-    this.service.deleteProduct(id).subscribe(() => {
-      this.loadProducts(); 
-    });
-  }
-
   editProduct(id: number) {
     this.router.navigate(['products/edit', id]);
   }
 
-  openDeleteDialog(id: number) {
-    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
-      data: { message: 'Deseja realmente excluir este produto?' },
-    });
-    dialogRef.afterClosed().subscribe((result) => {
-      if (result) {
-        this.deleteProduct(id);
-      }
-    });
+  deleteProduct(id: number) {
+    if (confirm('Deseja realmente excluir este produto?')) {
+      this.service.deleteProduct(id).subscribe(() => {
+        this.loadProducts(); 
+      });
+    }
   }
 
   goToHome() {
