+++ template/frontend/ambev.developer-evaluation/src/app/users/component/user-list/user-list.component.ts
@@ -1,8 +1,7 @@
 import { Component, inject, OnInit } from '@angular/core';
 import { User } from '../../models/user.model';
 import { CommonModule } from '@angular/common';
-import { Router, RouterOutlet } from '@angular/router';
-import { ConfirmationDialogComponent } from '../../../core/shared/components/confirmation-dialog/confirmation-dialog.component';
+import { Router } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
 import { MatIconModule } from '@angular/material/icon';
 import { MaterialModule } from '../../../core/shared/modules/material.module';
@@ -11,7 +10,7 @@ import { UserService } from '../../services/user.service';
 @Component({
   selector: 'app-user-list',
   standalone: true,
-  imports: [CommonModule, RouterOutlet, MaterialModule, MatIconModule],
+  imports: [CommonModule, MaterialModule, MatIconModule],
   templateUrl: './user-list.component.html',
   styleUrl: './user-list.component.scss',
 })
@@ -23,7 +22,7 @@ export class UserListComponent implements OnInit {
     'email',
     'firstname',
     'lastname',
-    'actions'
+    'actions',
   ];
 
   private dialog = inject(MatDialog);
@@ -44,25 +43,16 @@ export class UserListComponent implements OnInit {
     this.router.navigate(['users/add']);
   }
 
-  deleteUser(id: number) {
-    this.service.deleteUser(id).subscribe(() => {
-      this.loadUsers(); 
-    });
-  }
-
   editUser(id: number) {
     this.router.navigate(['users/edit', id]);
   }
 
-  openDeleteDialog(id: number) {
-    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
-      data: { message: 'Deseja realmente excluir este produto?' },
-    });
-    dialogRef.afterClosed().subscribe((result) => {
-      if (result) {
-        this.deleteUser(id);
-      }
-    });
+  deleteUser(id: number) {
+    if (confirm('Deseja realmente excluir este produto?')) {
+      this.service.deleteUser(id).subscribe(() => {
+        this.loadUsers();
+      });
+    }
   }
 
   goToHome() {
